[unit-testing] Add a seperate test for kinematic functions
