+++ unit-testings/CMakeLists.txt
@@ -21,6 +21,7 @@ add_so_test(test_acceleration_stabilization)
 add_so_test(test_model-base-ekf-flex-estimator-imu)
 add_so_test(test-kinetics-observer)
 add_so_test(test-dcm-bias-estimation)
+add_so_test(test-kinematics)
 
 add_custom_command(
   TARGET test_model-base-ekf-flex-estimator-imu
